python-client/requirements.txt
kafka-python
minio
Faker
